requirements.txt
torch
transformers
pandas
numpy
scikit-learn
tqdm
python-dotenv
google-generativeai
llama-index
llama-index-llms-gemini

